+++ NEWS.md
@@ -8,9 +8,9 @@
   now returns itself by default, instead of the parent `Idf` object. This
   enables to check the modified coordinate systems by printing the `IdfGeometry`
   (#506 #507).
-
 * Fix `IdfViewer$snapshot()` for PNG output. A new parameter `webshot` has been
   added in `IdfViewer$snapshot()` with default value being `FALSE`. (#509).
+* Now `IdfSchedule$... <- val` works (#512).
 
 # eplusr 0.15.1